+++ CHANGELOG.md
@@ -10,11 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SharePointDsc
   - Added automatic release with a new CI pipeline
   - Updated PULL_REQUEST_TEMPLATE.md to match DSC standard
+- SPFarm
+  - Added possibility to set application credential key.
+- SPTrustedSecurityTokenIssuer
+  - Fixed RegisteredIssuerNameRealm not applied if specified.
 
 ### Changed
 
 - SharePointDsc
-  - Updated all resources to load Utils module, which broke with the new CI
+  - Updated all resources and Invoke-SPDscCommand function to automatically
+    load Utils module, which broke with the new CI
 - SPConfigWizard
   - Updated checks in Set method to make sure the resource also runs when
     a language pack is installed